Set frame rate to 20 fps (#4)
Set the `frame_rate` to `20` fps. Previously, the trigger was set to
`40` fps, but it was later discovered that the camera was configured to
the wrong triggering GPIO, causing the camera FPS to operate at half the
defined value. The updated ROS camera driver now ensures the camera FPS
is consistent with the trigger frame rate, removing the need to
compensate by setting the frame rate to double the desired triggering
FPS.